+++ extensions/browser-bookmarks/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Browser Bookmarks Changelog
 
+## [Support for Dia and Ghost Browser] - {PR_MERGE_DATE}
+
+- Added support for `Dia` browser.
+- Added support for `Ghost Browser`.
+
 ## [Support Island Browser] - 2025-01-30
 
 - Added support for `Island` browser.